refactor: move training capability audit into operations
